requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
azure-search-documents==11.4.0
azure-core==1.29.5
python-dotenv==1.0.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.0
